pyproject.toml
[project]
name = "open-xtract"
version = "0.1.0"
description = "Extract structured data from documents, images, audio, and video using LLMs"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Cole McIntosh" }
]
keywords = ["llm", "extraction", "pydantic", "ai", "document", "structured-data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-ai-slim[google,logfire]>=1.37.0",
]

[project.urls]
Homepage = "https://github.com/colesmcintosh/open-xtract-v2"
Repository = "https://github.com/colesmcintosh/open-xtract-v2"
Issues = "https://github.com/colesmcintosh/open-xtract-v2/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/open_xtract"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["UP047"]  # TypeVar pattern is valid for pydantic-ai compatibility

[tool.pytest.ini_options]
testpaths = ["tests"]
